third_party/blink/web_tests/http/tests/devtools/runtime/runtime-callFunctionOn-expected.txt

Tests TestRunner.RuntimeAgent.callFunctionOn usages.


Running: testThis
3

Running: testArguments
{"a":1,"b":2}
{"a":1,"b":2}
{"c":1,"d":2}
4
undefined
